Fix issue with signout icon needing a text label
